Refactored parsing into it's own object
